Add another pattern for git: g()
I've been using gl, ga, gc etc for a while, and this flat pattern is creating a
lot of garbage in the command line. Based on @wfleming Dotfiles, I'm going to
give a try on using the pattern `g l` for everything now so subcommands don't
bloat the root path, and it adds a nice hierarchy.

This commit also cleans up pretty old stuff. I need to clean up other things
later.