Flushes history to $HISTFILE after every command
This changes $PROMPT_COMMAND so every time a new command is run, the current
command is added to the history file ($HISTFILE).

Without this, the history file is only updated after the session is closed. This
is not good in my case: I have a Tmux session with many panes that share the
same $HISTFILE dynamically, based on the session name. Once I run a command in
one pane, I want to be able to see it in `history` in another pane
instantaneously (so I can use Ctrl+R with fzf).