Fixes Ruby+gem installations in latest Ruby versions
For some reason, latest versions of Rub will install gems in $HOME/.gems and
that breaks how Rbenv expects it to work (with gems referenced from inside
Rbenv's directory).

There's a drawback with this approach is that now gems  are shared among
different Ruby versions and that some times breaks (e.g "incompatible version"),
but for now it's fine locally. I need to find a better solution later.