Skip vcs ignores in the_silver_searcher
I couldn't find a particular string in a project, which was clearly in code, in
master. It made no sense. Upon investigation, I realized the name of the folder
in which the class was in was called `workers/vendors`.

I occurred to me that, some times, `vendor` directories are excluded in ignore
files, such as `bundle/vendor`, `vendor` in Rails, and other frameworks.

Opening `.gitignore`, it had the line `vendor`, which I assumed
the_silver_searcher was taking as a pattern to exclude, not a literal directory
in the root directory (e.g `./vendor`).  Removing that line made the results
show up. I also tested this with another tool, `rg`.

This commit just makes `ag` (the_silver_searcher) completely ignore any VCS
ignore files (e.g `.gitignore`).

Eventually, I will have to add those ignores directly to `ag/ignore`.