Fixes Postgres causing segmentation fault on MacOS M1 (Ruby)
`export PGGSSENCMODE=disable` is added to the bash environment variables to
prevent segmentation faults when using Postgres with Rails on MacOS M1.

The root cause of the issue is related to the GSSAPI authentication mechanism
used by Postgres, which can lead to segmentation faults on certain systems,
especially when running Rails applications. By disabling GSSAPI encryption
mode with `PGGSSENCMODE=disable`, we can avoid these crashes and ensure
smooth operation of the Rails server.