Adds bin/alias/media_image_video.bashrc for resizing images with ImageMagick
Some times I want to share images from camera but the files are too large. This
script resizes images in a directory to a specified max width, preserving aspect
ratio and stripping EXIF data. It modifies files in place with no backup, so use
with caution.

That way I don't need to use any GUI tools, and can quickly resize a batch of
images from the command line.