bin/meeting_prep
This script reads through my notes and calendar to prepare pre-meeting notes
for the day. These notes are private and intended for my personal use.

**Calendar**
It reads from Apple calendar, so it might not always be up-to-date unless you
keep the app open. It requires icalBuddy.

**Notes**
It reads notes from my Obsidian vault (pure files) to try to find the person
associated with the meeting. The frontmatter of a particular person's note has
to include an `email:` property that matches the calendar meeting.

**Archeology on previous notes**
It reads daily notes up to a month ago (configurable), to find previous mentions
of that person. It looks for bullet points indented under a `[[Person Name]]`
with my previous notes.

Once that is done, it uses an LLM to summarize those notes into a concise summary
for quick reference. It includes stuff that I should follow-up on, or life updates
that I should remember.

**Format**
It outputs markdown notes that look like the following:

    - [[@John Doe]]
      - Previous discussions:
        - Discussed project X and Y.
        - Mentioned family vacation plans.
      - Follow-ups:
      - Agenda: